Added getPlugin and new param to loadView
The new param will probably be replaced later on. The idea being that
modules have their own template files when shipped (most of the time*)
so the developer specifies the name of the module template dir, checks
the current Theme for that template file, then falls back to the
original one that shipped.