Added Theme type to organize themes by accessbility
Themes come in 2 flavors: Site-wide themes, and Module-specific
templates. As of right now, this is the only way I know how to organize
it. For sake of not being confusing, the system has been given a Type
column where 0=themes and 1=Module. Module keeps it from being selected
in the Theme Manager.