Fixed an Installer issue
There was a bug a user had with Config.php not populating. Now we check
everytime for an existing config.php file AND to make sure it's
populated with something.|
The installer itself, has been given a manual fix for the problem, if
for whatever reason the Automatic didn't work, we hope to fix that by
allowing the admin to manually create the config.php file.