Expanded SQLError notification
There was no way to see if an error occured in the Installer, or even
during normal game operations due to how much info is displayed to the
client screen. Now we check if we're in the ezRPG installer and show the
same as SHOW_ERRORS, or just tell the client there's been an error.
Logging system should be created.