Fixed setMessage issue
header_msg hook was being called before at a modules initial startup. So
it always worked when you setMessage and then header(location)'d to a
new page since a new module would begin, activating the header_msg hook
again.
But if you were just going to loadView while in that current module
instance, then the message would be lost.

Now when a loadView is called, $this->getMessages() is called along with
it to do exactly what the hook did and fix the issue.