initial 0.0.1 Release of nuRPG
- Initial Release
- Modules are autoloaded with spl_autoload
- Support
for old modules still exist, but will be removed later
- Modules now are
names "MODULENAME.module.php" and are JUST put inside the modules/*
directory and nor a subdirectory of that.
- Module classes don't need
"Module_" prefixed to it. It's already in a module folder, extending a
Base_Module.
- Magic methods are to be created: "__activate",
"__deactivate", "__uninstall", and "__update". "__activate" will run
both install and activate functions, it'll be by design of the developer
to distinguish how these will be to avoid re-installation of DB
commands.
- Standard Formatting of PHPDoc comments will be utilized for
"Module Name", "Author", "Version" etc.
- An almost complete Plugin
Uploader that does away with the xml idea.
- Created an $app super
variable instead of using a proper IoC container to limit the number of
objects being passed by storing objects inside a single variable.
-
Currently we're pretty close to being 100% compatible with 1.2.1.x
modules. Some tweaks are needed for hooks and such.