0.0.2
- PluginManager.module.php now implements the activate, deactivate, and
uninstall functions (update to come later).
- PHPDoc comments are parsed
and utilized as part of the installer process.
- KillModuleCache has
been added to the PluginManager to address issues found during
development.
- Added a 'class' column to <ezrpg>plugins.
- Deleted
'filename', 'installed', and 'secondary_install' from <ezrpg>plugins.
-
'active' now takes the place of any 'installed' information. When a
plugin is uploaded, the system checks for __active() method. If exists,
it runs and checks for a True/False return. It's on the plugin
developers to implement their error tracking features to return correct
information. A FALSE return results in Active remaining as 0 (or NOT
installed) and TRUE results as 1 (installed and active). If a plugin
gets deactivated, then it's still installed, but the tables it installed
during it's initial activation should still be intact. The developer
must currently do a check to ensure tables aren't recreated. See
example.module.php for an example.